Refactor : 정적 팩토리 메서드 수정
Related:#80
Ref:#63,#42
