+++ application/src/main/java/core/application/movies/models/dto/response/MovieSearchRespDTO.java
@@ -1,15 +1,15 @@
-package core.application.movies.models.dto;
-
-import org.json.JSONObject;
+package core.application.movies.models.dto.response;
 
 import core.application.movies.models.entities.CachedMovieEntity;
 import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 
 @Data
 @Builder
 @Schema(description = "영화 검색 응답")
+@AllArgsConstructor
 public class MovieSearchRespDTO {
 
 	@Schema(description = "영화 ID", example = "A-12345")
@@ -23,15 +23,6 @@ public class MovieSearchRespDTO {
 
 	@Schema(description = "제작년도", example = "2015")
 	private String producedYear;
-
-	public static MovieSearchRespDTO from(JSONObject movie) {
-		return MovieSearchRespDTO.builder()
-			.movieId(movie.optString("movieId") + "-" + movie.optString("movieSeq"))
-			.title(movie.optString("title"))
-			.posterUrl(movie.optString("posters").split("\\|")[0])
-			.producedYear(movie.optString("prodYear"))
-			.build();
-	}
 
 	public static MovieSearchRespDTO from(CachedMovieEntity movie) {
 		return MovieSearchRespDTO.builder()
